Split out JSON logger from Container logger
This allows us to later reuse the JSON logging for other purposes and
detaches it from container specifics
