+++ lib/manageiq/loggers.rb
@@ -1,5 +1,8 @@
 require "manageiq/loggers/base"
+require "manageiq/loggers/json_logger"
+
 require "manageiq/loggers/cloud_watch"
 require "manageiq/loggers/container"
 require "manageiq/loggers/journald"
+
 require "manageiq/loggers/version"
